Integrate passage stage into workflow step
